Put a default commit element if no other commit is pointed at.
With this change if no specific commit is selected, the current head for
the repository is used, so that all the information is actually present
for every template and also consistent through all of them.

After this change, also convert all the templates so that commit_hash is
replaced by commit.sh1 and commit_desc by description or short_description
as needed.